+++ src/components/header/lib/authentication.ts
